+++ index.js
@@ -4,8 +4,6 @@ let fingerprint = require('./fingerprint')
 let getLambdaName = require('./get-lambda-name')
 let getLayers = require('./get-layers')
 let getRuntime = require('./get-runtime')
-let initArc = require('./init-arc')
-let initAWS = require('./init-aws')
 let initEnv = require('./init-env')
 let inventory = require('./inventory')
 let pathToUnix = require('./path-to-unix')
@@ -18,15 +16,13 @@ let validate = require('./validate')
 module.exports = {
   banner,         // Prints banner and loads basic env vars and AWS creds
   chars,          // Returns platform appropriate characters for CLI UI printing
-  fingerprint,    // Generates public/static.json for @static fingerprint true
+  fingerprint,    // Generates public/static.json for `@static fingerprint true`
   getLambdaName,  // Get Lambda name from Arc path
   getLayers,      // Get layer config from Arc file or config
   getRuntime,     // Get runtime config from Arc file or config
-  initArc,        // Initialize basic project configuration
-  initAWS,        // Initialize AWS env vars and credentials
   initEnv,        // Initialize env vars from .arc-env config
   inventory,      // Get inventory of current AWS resources from Arc file
-  pathToUnix,     // Use `/` seperated paths everywhere
+  pathToUnix,     // Normalize to `/` seperated paths for Windows-specific calls
   portInUse,      // Checks to see if a port is in use
   readArc,        // Reads Arc file and returns raw + parsed versions
   toLogicalID,    // Converts dash casing into Pascal casing for CloudFormation